backend/flask/requirements.txt
flask==3.0.2
flask-cors==4.0.0
pandas==2.2.1
openpyxl==3.1.2
langchain-community==0.0.20
chromadb==0.4.24
ollama==0.1.6
python-dotenv==1.0.0
requests==2.31.0
matplotlib==3.8.2
numpy==1.26.3
langchain-core==0.1.23
tqdm==4.66.1  # For progress bars in data processing
xlrd==2.0.1   # For legacy Excel file support